SWBIT-4144 #comment Converts HeaderRightContent and UserInfo in components
