+++ packages/frontend/src/components/EditDocument.tsx
@@ -8,6 +8,8 @@ import useUserData from "../lib/useUserData";
 import DocumentInfoForm from "./DocumentInfoForm";
 import EditTitleModal from "./EditTitleModal";
 import PublishBar from "./PublishBar";
+import HeaderRightContent from "./HeaderRightContent";
+import UserInfo from "./UserInfo";
 import {
   DOCUMENT_QUERY,
   CREATE_DOCUMENT_MUTATION,
@@ -105,10 +107,15 @@ const EditDocument: FC<EditDocumentProps> = ({ id, type }) => {
 
   if (loading) return <Loading color={documentType.color} />;
 
-  const { title, description, public: isPublic, example: isExample, advancedMode } =
-    document || {};
+  const {
+    title,
+    description,
+    public: isPublic,
+    example: isExample,
+    advancedMode
+  } = document || {};
 
-  window.sessionStorage.setItem('advancedMode', `${advancedMode}`);
+  window.sessionStorage.setItem("advancedMode", `${advancedMode}`);
 
   const location = window.location;
   const publicUrl = `${location.protocol}//${location.host}/app/public-document/${type}/${id}`;
@@ -191,6 +198,12 @@ const EditDocument: FC<EditDocumentProps> = ({ id, type }) => {
     </Document.Tab>
   );
 
+  const headerRightContent: Element = (
+    <HeaderRightContent>
+      <UserInfo name={user.name} />
+    </HeaderRightContent>
+  );
+
   return (
     <>
       <EditorComponent
@@ -216,6 +229,7 @@ const EditDocument: FC<EditDocumentProps> = ({ id, type }) => {
         }
         changeAdvancedMode={onSetAdvancedMode}
         documentAdvancedMode={advancedMode}
+        headerRightContent={headerRightContent}
       />
       {isEditTitleVisible && (
         <EditTitleModal
@@ -242,5 +256,5 @@ const Loading = styled(Spinner)<LoadingProps>`
   width: 100%;
   height: 100%;
   color: white;
-  background-color: ${props => props.color};
+  background-color: ${(props: LoadingProps) => props.color};
 `;
