+++ packages/frontend/src/components/HeaderRightContent.tsx
@@ -1,0 +1,13 @@
+import React, { FC } from "react";
+import styled from "@emotion/styled";
+
+const HeaderRightContent = styled.div`
+  align-items: center;
+  display: flex;
+  height: 70px;
+  position: absolute;
+  right: 20px;
+  top: 0;
+`;
+
+export default HeaderRightContent;
